Do not force spaces before inserted block when pasting behind eol
When pasting a block behind the end of certain lines an the pasted
block line starts with whitespaces, we just move to the first
non-whitespace character to be yanked. This may not insert spaces but
tabs. Note that this is impossible if we do not paste behind eol
because then the insertion of spaces is required to shift the text
behind the insertion point forward.