Fix linewise `evil-delete' and `evil-change' at the last line in buffer
This case has to be handled differently because the last line does not
end in a newline. Therefore using `delete-region' would to remove the
last line of the buffer (just makes it empty). In this case the
*previous* newline has to be deleted.

Similary `evil-change' has to use `evil-insert-below' instead of
`evil-insert-above' after deleting the last line.