Add `evil-loop' and `evil-motion-loop' macros
These patterns turn up everywhere in the motion code (and even in
Emacs' own movement commands as well: cf. the wasteful verbosity
of `forward-sentence' in paragraphs.el).

`evil-loop' provides a countdown to zero, allowing negative counts.
`evil-motion-loop' is similar, but quits when point does not move.
This makes it easy to honor point #6 in the requirements, that the
return value is the number of unsuccessful iterations.