Skip block-insertion on lines after end-of-line
When calling `evil-insert' from visual-block mode, the inserted text
should not be inserted on lines on which the insertion point is behind
the end of that line (in contrast to `evil-append'). For this an
additional parameter `skip-empty-lines' has been added to
`evil-insert' and `evil-append' and is handled in the interactive
forms for visual mode.