Add input method support
Add an :input-method keyword argument to `evil-define-state' for
specifying whether input methods are enabled or disabled in the state.
When an input method is disabled, it is saved in `evil-input-method'
so that a future state may re-enable it. Insert state and Emacs state
enable input methods, while the other states disable them.