Fix "C" and "D" in Visual state
When `evil-delete-line' ("D") is invoked on a characterwise Visual
selection, it acts linewise. When the selection is blockwise, it
deletes to the end of each line in the block. `evil-change-line' ("C")
works similarly, but calls `evil-change' to enter Insert state
intelligently. Add a DELETE-FUNC argument to `evil-change' to make
it more reusable.