Do not record ESC for repeating in normal-state
The command `evil-normal-state' is recorded using keystrokes which is
required when insert-state is exited. Therefore ESC should not be
bound to the same command in normal-state because this would record a
literal ESC in the repeat ring. Now, ESC is bound to
`evil-force-normal-state' in normal-state, which does the same as
`evil-normal-state' but without recording anything in the repeat-ring.