Add "+ and "* registers
"+ accesses the system clipboard (modified by Ctrl+C and Ctrl+V in
other programs). "* accesses the primary X11 selection (made with
the mouse, not available on Windows).

Note that Emacs provides the variables `x-select-enable-clipboard'
and `x-select-enable-primary' which, when set to t, enable
integration with the above. Then a plain yank ("yy") or paste ("P")
automatically accesses the clipboard and/or primary selection.