Substitute keymap alists for ordered lists of keymaps
Make `evil-state-keymaps' and its dependencies return keymap alists,
which are joined by the utility function `evil-concat-keymap-alists'.
A keymap alist is understood to be an ordered list of cons cells
(VAR . MAP) where each keymap may occur only once, but where the
variables may be repeated. This allows `evil-mode-map-alist' to reuse
variables found in `minor-mode-map-alist', so that deactivating
keymaps in the latter immediately deactivates related keymaps in
the former.