Fix bug with the initial state after 5c56bad
For Evil to work properly with meta-modes like noweb and MuMaMo,
Evil should be initialized only once in a given buffer. However,
the major mode of the buffer may not have been set up when
`evil-local-mode' is called for the first time. We solve this
by postponing Evil's initialization with `post-command-hook'.