Disable `evil-esc-mode' during a call to `key-binding' (addresses #63)
In Emacs 24 a "bug" of the function `key-binding' has been fixed: if a
direct binding to "ESC" is available in an active keymap this binding
is always returned even if a longer key-binding, e.g., an ESC-sequence
is available in another keymap. The former behavior was to return the
longer match. Because a direct binding to ESC is active in evil's
`evil-esc-map' keymap, which is used to multiplex the ESC key in
terminal mode, determining the active binding of ESC-sequences does
not work anymore. This patch temporarily disables `evil-esc-mode' and
the corresponding keymap `evil-esc-map' during a call to
`key-binding'. This should not invalidate the original purpose of this
keymap because the global key reader does not depend on the function
`key-binding'.