Ensure test buffer is not empty in some tests
Some tests of the repeat system involving operators formerly worked on
empty buffers. This caused the motions to raise an error. The new more
restrictive handling of error in operator state prevented the
corresponding command to be recorded for repeating. This patch ensures
that the test buffer is not empty and thus no error is raised.