Ensure `evil-forward-word-begin' does not move back beyond original point
If a line starts with whitespaces and point is placed at one of them
the operator motion `evil-forward-word-begin' should not move point
back to the previous line. This patch ensures that a move-back only
happens if the starting point of the motion is on a previous line.
This fixes issue #84