Make `evil-esc' hide itself from `last-command'
`evil-esc' is a special command only used to handle the ESC key
correctly in terminals. Some commands, e.g., org mode's `org-cycle'
depend on inspecting the variable `last-command'. This patch sets
`this-command' to `last-command' in `evil-esc' which gives the correct
value of `last-command' in any command called via an ESC key sequence.