Uninterned symbol in `evil-define-key', `evil-execute-in-normal-state'
The uninterned symbol works as a placeholder for an anonymous
function, letting the function refer to itself for the purpose
of removing itself from a hook. This obviates the need for
`evil-generate-symbol'. Furthermore, `evil-old-move-cursor-back'
is also unnecessary since the old value can be spliced into the
lambda expression (a poor man's closure).