Handle single ESC in operator state more accurately (fixes #89)
Operator state is special because it reads the key sequence for the
motion itself using `read-key'. This patch multiplexes the ESC key
similar to the `evil-esc' when the motion key-sequence is read in
`evil-keypress-parser'. Furthermore [escape] is bound to
'keyboard-quit in operator state and keyboard-quit now aborts operator
state completely. This means that the global variables describing the
operator motion are not set (they are not used anyway, because no
motion has been read and no operator is about being executed) and the
repeat recording is aborted.