Replace `evil-window-lru' with `evil-window-mru'
This function selects the buffer that is different from the current
buffer and is shown in another window. This behavior is close to, but
not identical, Vim's behavior for the C-W p command, that selects the
previously selected window. It differs from Vim if the current buffer
is shown in multiple windows. This addresses issue #127.