Fix `evil-eobp' at second last character of last line
In this case the call to `forward-word' causes point to be moved to
the last character of the last line, which is not eob. The former
implementation did another call to `forward-word' in this situation.
This must only happen if we point is at the final newline after the
first `forward-word'. This fixes issue #153.