Modify `evil-define-text-object' calling body only once
This changes the interface of evil-define-text-object. The body of a
text object is now called exactly once (when called interactively).
Furthermore the framework code of `evil-define-text-object' assumes
that the body handles the extension of the current visual selection on
its own. The current implementation of most text objects does *not* do
this; this has to be fixed in subsequent changes.