Add mouse-2 binding to insert state
In Emacs 23 we have to bind mouse-2 to `mouse-yank-primary' in insert
state, too. By default it's bound to `mouse-yank-at-click' which is
effectively a `yank' function and not an insert function. The problem
with this function is that it pretends itself being a `yank' command.
On the one hand this confuses the repeat system (changes
`this-command') and on the other hand a repeated `yank' is differently
from an insert command (always inserts the actual content of the
kill-ring).