Fix visual `evil-apply-on-rectangle' after line-movement
If this function is called after a line-movement `next-line' or
`previous-line' it takes `temporary goal column into account'. The
column this value corresponds to is the column in which point should
be placed. But because the selection is *exclusive* we have to use
this value + 1 to find the correct end column of the selection.
This fixes issue #199.