Set initial state in `evil-initialize'.
The function `evil-initialize' is called by
`define-globalized-minor-mode' whenever evil should be started in a
new buffer or if the major mode in a buffer has been changed. In
particular, `define-globalized-minor-mode' calls this function exactly
once in each buffer unless the major-mode has been changed. This is
ensured using major-mode-hooks as well as post-command-hooks.

Thus `evil-initialize' is the perfect candidate for setting the
initial evil in a buffer. The old implementation adviced
`evil-mode-check-buffers', which is defined by
`define-globalized-minor-mode', to setup the initial state.
Unfortunately this advice did not check whether the initialization for
the current major-mode has already been performed. In some cases this
caused a double-initialization of this state when the first command in
a buffer is executed. The reason is that `evil-mode-check-buffers' is
called from a post-command-hook which may not be called before the
first command is being executed. But if this command changed the state
(e.g. "i" changing to insert state), the double-initialization caused
evil to switch back to the initial state immediately.

This fixes #172.