Handle extension to eol in `evil-apply-on-block`.
`evil-apply-on-block` did not handle the case of a visual extension to the eol correctly.

This change unifies `evil-apply-on-block` and
`evil-apply-on-rectangle` into a single function. The only difference
between two is that the former should pass buffer positions the latter
column numbers to the function to be applied on the block. This can
now be chosen by a parameter of `evil-apply-on-block`.
`evil-apply-on-rectangle` is now a simple convenience wrapper always
passing the columns.