Make `evil-jump-item` jump out of a string if appropriate.
`evil-jump-item` is syntax sensitive: if (point) is not in a
string/comment only the string/comment content is considered, if
(point) is not in a string or comment, strings and comments are
skipped.

In one situation this may not be desired: if (point) is within a
string but no jump-target is found within the string. Then the old behavior would not move (point) at all. This is a common case in programming modes, e.g.

   printf("some text");

If point is within the string literal, then `evil-jump-point` should
move to the opening parenthesis.

The new behavior of `evil-jump-item` is now to look after the end of
the string as if it has been called right after the end of string, if
no appropriate jump target could be found within the string.

This addresses issue #249.