Update X selection in pre-command hook (fix #290)
If the X selection is updated in post-command hook and the command has
been a deletion, then the text to be put in the X selection is not
correct. In particular, if this text is empty then the X selection
would be updated with an empty string which may cause strange side
effects.