Make `*` and `#` search for symbols optionally (fix #360)
The new custom variable `evil-symbol-word-search` determines whether
the word search function family (bound to `*` and `#`) looks for the
symbol (if non-nil) or word (if nil) under point. The variable is
buffer-local so it can be set appropriately for each major-mode. The
default value is `nil` to mimic Vim's default behaviour.