implement `evil-ex-echo` directly using overlays (fix #459)
The old implementation of `evil-ex-echo` used `minibuffer-message`.
Unfortunately this function calls `sit-for` which in turn used
`unread-command-events`. This causes the Emacs macro recording system
to record some events more than once if `evil-ex-echo` is called
during the recording process. This change implements the ex echo
message without `sit-for` but uses overlays and a pre-command-hook
instead.